Added class - component
