+++ package.json
@@ -5,9 +5,9 @@
   "description": "Личный проект «Большое путешествие» от HTML Academy",
   "scripts": {
     "eslint": "eslint src/**/*.js",
-    "test": "npm run eslint",
     "build": "webpack",
-    "start": "webpack-dev-server --open"
+    "start": "webpack-dev-server --open",
+    "test": "npm run eslint"
   },
   "devDependencies": {
     "@babel/core": "7.3.4",
