+++ src/utils.js
@@ -10,7 +10,7 @@ export const clearElement = (element) => {
 };
 
 
-export const createContainer = (template) => {
+export const createElement = (template) => {
   const newElement = document.createElement(`div`);
   newElement.innerHTML = template;
   return newElement.firstChild;
